Force package-id to be integer

+++ ote/src/clj/ote/services/transit_changes.sql
@@ -53,15 +53,15 @@ SELECT ts.id AS "transport-service-id",
  ORDER BY "change-date" ASC, "interfaces-has-errors?" DESC, "no-interfaces?" DESC, "no-interfaces-imported?" ASC;
 
 -- name: calculate-routes-route-hashes-using-headsign
-SELECT calculate_route_hash_id_using_headsign(:package-id);
+SELECT calculate_route_hash_id_using_headsign(:package-id::INTEGER);
 -- name: calculate-routes-route-hashes-using-short-and-long
-SELECT calculate_route_hash_id_using_short_long(:package-id);
+SELECT calculate_route_hash_id_using_short_long(:package-id::INTEGER);
 -- name: calculate-routes-route-hashes-using-route-id
-SELECT calculate_route_hash_id_using_route_id(:package-id);
+SELECT calculate_route_hash_id_using_route_id(:package-id::INTEGER);
 -- name: calculate-routes-route-hashes-using-long-headsign
-SELECT calculate_route_hash_id_using_long_headsign(:package-id);
+SELECT calculate_route_hash_id_using_long_headsign(:package-id::INTEGER);
 -- name: calculate-routes-route-hashes-using-long
-SELECT calculate_route_hash_id_using_long(:package-id);
+SELECT calculate_route_hash_id_using_long(:package-id::INTEGER);
 
 -- name: fetch-services-with-route-hash-id
 SELECT ts.id as "service-id", ts.name as service, op.name as operator, d."route-hash-id-type" as type
